Use FZF in Vim and not Ctrl-P
The fact that I was using Ctrl-P instead of FZF was an artifact of a
different, less-efficient, era of my dotfiles. Now that I've been
working on them I've decided that it's time to bite the bullet and move
to FZF. This branch contains all of the work of replacing Ctrl-P and
configuring FZF so that it fits my workflow.