Fix Showing Git Status in Airline
I've been having issues with the branch and git status being displayed
in my status line, and it turns out that the match-count-statusline
plugin was hard overwriting the airline_section_b variable, which made
it so that the git status could not be displayed.

I could write a custom after directory vim file to fix this, but I'm a
little too lazy, and I'm not that heartbroken about losing this one
plugin.