Just gonna send it!
![](https://media.tenor.com/images/c5abe9e755d832eb6e53050311f2de65/tenor.gif)

Hello future Patrick. This was a day where you once again gave up all
fucks and decided to change how your entire Vim setup works. Here you've
fully ripped out ALE and decided to go with coc.nvim because it has
_much_ better Python language server support.

I don't know how well this is going to go, but here we are.

This PR is going to deprecate the following PRs as it has merged both of
them and then added more things:
* https://github.com/mr-mustash/dotfiles/pull/35
* https://github.com/mr-mustash/dotfiles/pull/36

Good luck future Patrick. You're going to need it.